fix: split collection FAQ data from client component to fix server import (fixes #2532)
